Revert "Rename hotel management app"
