+++ hotel-management-app-v1/src/main/java/com/fb/hotelmanagementappv1/models/City.java
@@ -1,4 +1,4 @@
-package com.fb.hotelApp.models;
+package com.fb.hotelmanagementappv1.models;
 
 public class City {
     private int id;
