+++ hotel-management-app-v1/src/main/java/com/fb/hotelmanagementappv1/models/Hotel.java
@@ -1,4 +1,4 @@
-package com.fb.hotelApp.models;
+package com.fb.hotelmanagementappv1.models;
 
 public class Hotel {
     private int id;
