+++ hotel-management-app-v1/src/test/java/com/fb/hotelmanagementappv1/HotelManagementAppV1ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.fb.hotelApp;
+package com.fb.hotelmanagementappv1;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
